# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for Barber Cukur-in",
        "title": "Barber Cukur-in",
        "contact": {
            "name": "Nuryanto",
            "url": "https://www.linkedin.com/in/nuryanto-1b2721156/",
            "email": "nuryantofattih@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/barber-service/barber/auth/change_password": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Change Password",
                "parameters": [
                    {
                        "type": "string",
                        "description": "OS Device",
                        "name": "OS",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "OS Device",
                        "name": "Version",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "req param #changes are possible to adjust the form of the registration form from frontend",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ResetPasswd"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/tool.ResponseModel"
                        }
                    }
                }
            }
        },
        "/barber-service/barber/auth/forgot": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Forgot Password",
                "parameters": [
                    {
                        "type": "string",
                        "description": "OS Device",
                        "name": "OS",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "OS Device",
                        "name": "Version",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "req param #changes are possible to adjust the form of the registration form from frontend",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ForgotForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/tool.ResponseModel"
                        }
                    }
                }
            }
        },
        "/barber-service/barber/auth/login": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "type": "string",
                        "description": "OS Device",
                        "name": "OS",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "OS Device",
                        "name": "Version",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "req param #changes are possible to adjust the form of the registration form from frontend",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.LoginForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/tool.ResponseModel"
                        }
                    }
                }
            }
        },
        "/barber-service/barber/auth/logout": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "logout",
                "parameters": [
                    {
                        "type": "string",
                        "description": "OS Device",
                        "name": "OS",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "OS Device",
                        "name": "Version",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/tool.ResponseModel"
                        }
                    }
                }
            }
        },
        "/barber-service/barber/auth/register": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Register Owner",
                "parameters": [
                    {
                        "type": "string",
                        "description": "OS Device",
                        "name": "OS",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "OS Device",
                        "name": "Version",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "req param #changes are possible to adjust the form of the registration form from frontend",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.RegisterForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/tool.ResponseModel"
                        }
                    }
                }
            }
        },
        "/barber-service/barber/auth/verify": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Verify",
                "parameters": [
                    {
                        "type": "string",
                        "description": "OS Device",
                        "name": "OS",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "OS Device",
                        "name": "Version",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "req param #changes are possible to adjust the form of the registration form from frontend",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.VerifyForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/tool.ResponseModel"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.ForgotForm": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "string"
                }
            }
        },
        "models.LoginForm": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "string"
                },
                "fcm_token": {
                    "type": "string"
                },
                "pwd": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "models.RegisterForm": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
        "models.ResetPasswd": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "string"
                },
                "confirm_pwd": {
                    "type": "string"
                },
                "pwd": {
                    "type": "string"
                }
            }
        },
...
        "models.VerifyForm": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "string"
                },
                "verify_code": {
                    "type": "string"
                }
            }
        },
        "tool.ResponseModel": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "message": {
                    "description": "Code int         `json:\"code\"`",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}